requirements.txt
pytest>=7.4
playwright>=1.43
allure-pytest>=2.13.2
